Fix a windows warning for type conversion.
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@719143 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/commands/Message.cpp
@@ -460,8 +460,8 @@ unsigned int Message::getSize() const {
 
     unsigned int size = DEFAULT_MESSAGE_SIZE;
 
-    size += this->getContent().size();
-    size += this->getMarshalledProperties().size();
+    size += (unsigned int)this->getContent().size();
+    size += (unsigned int)this->getMarshalledProperties().size();
 
     return size;
 }
